Move FileType to org.gradle.api.file
